+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2025-12-23
+
+### Added
+- Projectile grabbing functionality
+
+### Changed
+- Reorganized Risk of Options settings, sorry you might need check your old config to port it over
+
 ## [1.4.6] - 2025-12-22
 
 ### Removed
@@ -12,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Simplified persistence logic to always only capture bagged objects
-- Persistence now only applies to objects with dither models (like common chests and shrines)
+- Persistence now works with objects with dither models (like common chests and shrines)
 
 ### Fixed
 - Multiplayer synchronization for removing objects from persistence on impact